separate rendering functions into separate module
